Rename loop file
